More experimental results out of ResultTest.cpp
